+++ main/Ultrasonic.h
@@ -7,7 +7,7 @@ class Ultrasonic {
 	public:
 		Ultrasonic();
 		Ultrasonic(int trig_pin, int echo_pin);
-		float getDistance();
+		double getDistance();
     private:
 		int m_trig_pin;
 		int m_echo_pin;
